requirements.txt
fastapi
scikit-learn
numpy
pandas
requests
python-dotenv
uvicorn
matplotlib
plotly
umap-learn
tqdm 